Preemptive bugfixing part 3

+++ Server/src/main/java/core/game/node/entity/skill/gather/mining/MiningNode.java
@@ -636,7 +636,7 @@ public enum MiningNode{
         public int getRespawnDuration() {
                 int minimum = respawnRate & 0xFFFF;
                 int maximum = (respawnRate >> 16) & 0xFFFF;
-                double playerRatio = ServerConstants.MAX_PLAYERS / Repository.getPlayers().size();
+                double playerRatio = (double) ServerConstants.MAX_PLAYERS / Repository.getPlayers().size();
                 return (int) (minimum + ((maximum - minimum) / playerRatio));
         }
 }
